Rename studio/api/ -> studio/queries/
