+++ sanity.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "sanity";
 import { structureTool } from "sanity/structure";
 import { schemaTypes } from "@/studio/schemas";
-import { dataset, projectId } from "@/studio/api/client";
+import { dataset, projectId } from "@/studio/queries/client";
 import { structure } from "@/studio/sanity-structure";
 
 export default defineConfig({
